+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Apply remotion rate limit on a per-user basis
+
 ## [0.11.2] - 2022-03-25
 
 ### Fixed